+++ app/src/mspsp-qt-app/activity.cpp
@@ -3,3 +3,22 @@
 Activity::Activity()
 {
 }
+
+Activity::Activity(int id, std::string name, int duration)
+{
+    this->SetActivitySettings(id, name, duration);
+}
+
+void Activity::SetActivitySettings(int id, std::string name, int duration)
+{
+    this->id = id;
+    this->name = name;
+    this->duration = duration;
+    this->startingTime = 0;
+    this->finishingTime = 0;
+}
+
+int Activity::GetDuration()
+{
+    return this->duration;
+}