+++ app/src/mspsp-qt-app/activity.h
@@ -1,6 +1,10 @@
 #ifndef ACTIVITY_H
 #define ACTIVITY_H
 
+#include <string>
+#include <vector>
+#include <algorithm>
+
 /*
     This class is a representation of an activity in a project.
 */
@@ -8,10 +12,13 @@ class Activity
 {
 public:
     Activity();
+    Activity(int id, std::string name, int duration);
+    void SetActivitySettings(int id, std::string name, int duration);
+    int GetDuration();
 private:
     int id;
-    string name;
-    double duration;
+    std::string name;
+    int duration;
     int startingTime;
     int finishingTime;
 };