+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+0.0.4    2013-04-21
+	- Added tests project for unit tests.
+	- Added unit tests for Project and Activity classes.
+	- Fixed missing libraries in Project and Activity classes.
+
+
 0.0.3    2013-04-19
 
     - Added Project and Activity classes.